+++ src/app/services/mocks/onedata-graph.js
@@ -202,7 +202,38 @@ const tokenHandlers = {
       } else {
         return {
           success: false,
-          error: { id: 'badToken' },
+          error: { id: 'badValueToken' },
+          data: {},
+        };
+      }
+    } else {
+      return messageNotSupported;
+    }
+  },
+  verify_invite_token(operation, entityId, data) {
+    if (operation === 'create') {
+      const token = get(data, 'token');
+      const tokenRecord = this.get('store').peekAll('token').findBy('token', token);
+      if (tokenRecord.get('revoked')) {
+        return {
+          success: false,
+          error: { id: 'tokenRevoked' },
+          data: {},
+        };
+      } else if (tokenRecord.get('caveats').length > 0) {
+        return {
+          success: false,
+          error: {
+            id: 'tokenCaveatUnverified',
+            details: {
+              caveat: tokenRecord.get('caveats')[0],
+            },
+          },
+          data: {},
+        };
+      } else {
+        return {
+          success: true,
           data: {},
         };
       }